Rename `images` to `gdstore`
