+++ daemon/containerd/image_delete.go
@@ -13,7 +13,7 @@ import (
 	"github.com/docker/docker/api/types/events"
 	imagetypes "github.com/docker/docker/api/types/image"
 	"github.com/docker/docker/container"
-	dimages "github.com/docker/docker/daemon/images"
+	"github.com/docker/docker/daemon/images/gdstore"
 	"github.com/docker/docker/image"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/opencontainers/go-digest"
@@ -57,7 +57,7 @@ func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, 
 	start := time.Now()
 	defer func() {
 		if retErr == nil {
-			dimages.ImageActions.WithValues("delete").UpdateSince(start)
+			gdstore.ImageActions.WithValues("delete").UpdateSince(start)
 		}
 	}()
 
@@ -75,7 +75,7 @@ func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, 
 	if img == nil {
 		if len(all) == 0 {
 			parsed, _ := reference.ParseAnyReference(imageRef)
-			return nil, dimages.ErrImageDoesNotExist{Ref: parsed}
+			return nil, gdstore.ErrImageDoesNotExist{Ref: parsed}
 		}
 		imgID = image.ID(all[0].Target.Digest)
 		var named reference.Named
@@ -90,7 +90,7 @@ func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, 
 		}
 
 		if len(sameRef) == 0 && named != nil {
-			return nil, dimages.ErrImageDoesNotExist{Ref: named}
+			return nil, gdstore.ErrImageDoesNotExist{Ref: named}
 		}
 
 		if len(sameRef) == len(all) && !force {
