+++ daemon/create.go
@@ -14,7 +14,7 @@ import (
 	networktypes "github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/container"
 	"github.com/docker/docker/daemon/config"
-	"github.com/docker/docker/daemon/images"
+	"github.com/docker/docker/daemon/images/gdstore"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/image"
 	"github.com/docker/docker/internal/multierror"
@@ -89,7 +89,7 @@ func (daemon *Daemon) containerCreate(ctx context.Context, daemonCfg *configStor
 				Variant:      img.Variant,
 			}
 
-			if !images.OnlyPlatformWithFallback(p).Match(imgPlat) {
+			if !gdstore.OnlyPlatformWithFallback(p).Match(imgPlat) {
 				warnings = append(warnings, fmt.Sprintf("The requested image's platform (%s) does not match the detected host platform (%s) and no specific platform was requested", platforms.Format(imgPlat), platforms.Format(p)))
 			}
 		}
