+++ daemon/images/gdstore/image_changes.go
@@ -1,4 +1,4 @@
-package images
+package gdstore
 
 import (
 	"context"
