+++ daemon/images/gdstore/image_windows.go
@@ -1,4 +1,4 @@
-package images
+package gdstore
 
 import (
 	"context"
