+++ daemon/images/gdstore/store.go
@@ -1,4 +1,4 @@
-package images
+package gdstore
 
 import (
 	"context"
