+++ daemon/images/gdstore/store_test.go
@@ -1,4 +1,4 @@
-package images
+package gdstore
 
 import (
 	"context"
